[refactor]: Change useFetch to customFetch
